Improved the OpenTK rendering method
Update shaders to not use the (layout = ) attribute, but rather get the attribute locations in code.
Included the shader source as text files next to the project, so that they can easily be edited. They must still be embedded in the source code however.

This has reduced the minimum required OpenGL version to 3.0, which will improve compatability.